refactor: update imports to use type-only imports in adapter.ts
